refactor: address code review findings from PR #35
- Remove dead `yieldFocusIfNeeded()` method and `isSwitcherAppActive`
  dependency from AppSwitcher — activating the target app is sufficient
  to yield focus, so the guard body was empty and unused
- Remove dead `hideHUDApp` dependency from HUDManager and the test
  assertion that trivially passed because nothing called it
- Extract duplicate `MockWindow` NSWindow subclass into a shared
  TestHelpers.swift instead of defining it separately in
  PressAndHoldTests and SettingsWindowPresentationStateTests
- Rename SettingsWindowPresentationState.swift →
  SettingsWindowLifecycleCoordinator.swift to match the primary exported
  type; update xcodeproj and rename the corresponding test file
- Remove unused "Create your first group here..." localization key from
  all 15 .lproj files — the string was added to the welcome banner but
  not referenced in WelcomeBannerView
- Add comments on the two `lazy var` dependencies in HUDManager
  explaining why `lazy` is required (default captures `self`)
