+++ ShortcutCycle/ShortcutCycle/Resources/ar.lproj/Localizable.strings
@@ -147,4 +147,3 @@
 "Show welcome again" = "عرض الترحيب مجدداً";
 "ShortcutCycle is running in your menu bar" = "ShortcutCycle يعمل في شريط القوائم";
 "Look for the ShortcutCycle icon in the top-right menu bar whenever you want to open the app again." = "ابحث عن أيقونة ShortcutCycle في شريط القوائم أعلى اليمين عندما تريد فتح التطبيق مجدداً.";
-"Create your first group here, or turn on Open at Login so ShortcutCycle is always ready." = "أنشئ مجموعتك الأولى هنا، أو فعّل «فتح عند تسجيل الدخول» ليكون ShortcutCycle جاهزاً دائماً.";
