+++ ShortcutCycle/ShortcutCycle/Resources/de.lproj/Localizable.strings
@@ -146,4 +146,3 @@
 "Show welcome again" = "Willkommen erneut anzeigen";
 "ShortcutCycle is running in your menu bar" = "ShortcutCycle läuft in Ihrer Menüleiste";
 "Look for the ShortcutCycle icon in the top-right menu bar whenever you want to open the app again." = "Suchen Sie das ShortcutCycle-Symbol in der Menüleiste oben rechts, wenn Sie die App erneut öffnen möchten.";
-"Create your first group here, or turn on Open at Login so ShortcutCycle is always ready." = "Erstellen Sie hier Ihre erste Gruppe oder aktivieren Sie \u201eBeim Anmelden \u00f6ffnen\u201c, damit ShortcutCycle immer bereit ist.";
