+++ ShortcutCycle/ShortcutCycle/Resources/en.lproj/Localizable.strings
@@ -148,4 +148,3 @@
 "Show welcome again" = "Show welcome again";
 "ShortcutCycle is running in your menu bar" = "ShortcutCycle is running in your menu bar";
 "Look for the ShortcutCycle icon in the top-right menu bar whenever you want to open the app again." = "Look for the ShortcutCycle icon in the top-right menu bar whenever you want to open the app again.";
-"Create your first group here, or turn on Open at Login so ShortcutCycle is always ready." = "Create your first group here, or turn on Open at Login so ShortcutCycle is always ready.";
