+++ ShortcutCycle/ShortcutCycle/Resources/es.lproj/Localizable.strings
@@ -146,4 +146,3 @@
 "Show welcome again" = "Mostrar bienvenida de nuevo";
 "ShortcutCycle is running in your menu bar" = "ShortcutCycle está activo en tu barra de menús";
 "Look for the ShortcutCycle icon in the top-right menu bar whenever you want to open the app again." = "Busca el icono de ShortcutCycle en la barra de menús superior derecha cuando quieras abrir la aplicación de nuevo.";
-"Create your first group here, or turn on Open at Login so ShortcutCycle is always ready." = "Crea tu primer grupo aquí o activa «Abrir al iniciar sesión» para que ShortcutCycle esté siempre listo.";
