+++ ShortcutCycle/ShortcutCycle/Resources/fr.lproj/Localizable.strings
@@ -146,4 +146,3 @@
 "Show welcome again" = "Afficher à nouveau le message de bienvenue";
 "ShortcutCycle is running in your menu bar" = "ShortcutCycle est actif dans votre barre des menus";
 "Look for the ShortcutCycle icon in the top-right menu bar whenever you want to open the app again." = "Recherchez l'icône ShortcutCycle dans la barre des menus en haut à droite lorsque vous souhaitez rouvrir l'application.";
-"Create your first group here, or turn on Open at Login so ShortcutCycle is always ready." = "Créez votre premier groupe ici ou activez « Ouvrir à la connexion » pour que ShortcutCycle soit toujours disponible.";
