+++ ShortcutCycle/ShortcutCycle/Resources/it.lproj/Localizable.strings
@@ -146,4 +146,3 @@
 "Show welcome again" = "Mostra di nuovo il benvenuto";
 "ShortcutCycle is running in your menu bar" = "ShortcutCycle è attivo nella barra dei menu";
 "Look for the ShortcutCycle icon in the top-right menu bar whenever you want to open the app again." = "Cerca l'icona di ShortcutCycle nella barra dei menu in alto a destra quando vuoi riaprire l'app.";
-"Create your first group here, or turn on Open at Login so ShortcutCycle is always ready." = "Crea il tuo primo gruppo qui o attiva «Apri all'accesso» affinché ShortcutCycle sia sempre pronto.";
