+++ ShortcutCycle/ShortcutCycle/Resources/nl.lproj/Localizable.strings
@@ -146,4 +146,3 @@
 "Show welcome again" = "Welkomst opnieuw tonen";
 "ShortcutCycle is running in your menu bar" = "ShortcutCycle wordt uitgevoerd in de menubalk";
 "Look for the ShortcutCycle icon in the top-right menu bar whenever you want to open the app again." = "Zoek het ShortcutCycle-pictogram in de menubalk rechtsboven wanneer u de app opnieuw wilt openen.";
-"Create your first group here, or turn on Open at Login so ShortcutCycle is always ready." = "Maak hier uw eerste groep aan of schakel 'Openen bij aanmelding' in zodat ShortcutCycle altijd klaar staat.";
