+++ ShortcutCycle/ShortcutCycle/Resources/pl.lproj/Localizable.strings
@@ -146,4 +146,3 @@
 "Show welcome again" = "Pokaż ponownie powitanie";
 "ShortcutCycle is running in your menu bar" = "ShortcutCycle działa na pasku menu";
 "Look for the ShortcutCycle icon in the top-right menu bar whenever you want to open the app again." = "Szukaj ikony ShortcutCycle na pasku menu w prawym górnym rogu, gdy chcesz ponownie otworzyć aplikację.";
-"Create your first group here, or turn on Open at Login so ShortcutCycle is always ready." = "Utw\u00f3rz tutaj pierwsz\u0105 grup\u0119 lub w\u0142\u0105cz \u201eOtwieraj przy logowaniu\u201c, aby ShortcutCycle by\u0142 zawsze gotowy.";
