+++ ShortcutCycle/ShortcutCycle/Resources/pt-BR.lproj/Localizable.strings
@@ -146,4 +146,3 @@
 "Show welcome again" = "Mostrar boas-vindas novamente";
 "ShortcutCycle is running in your menu bar" = "ShortcutCycle está em execução na sua barra de menus";
 "Look for the ShortcutCycle icon in the top-right menu bar whenever you want to open the app again." = "Procure o ícone do ShortcutCycle na barra de menus no canto superior direito sempre que quiser abrir o aplicativo novamente.";
-"Create your first group here, or turn on Open at Login so ShortcutCycle is always ready." = "Crie seu primeiro grupo aqui ou ative «Abrir no Login» para que o ShortcutCycle esteja sempre pronto.";
