+++ ShortcutCycle/ShortcutCycle/Resources/ru.lproj/Localizable.strings
@@ -146,4 +146,3 @@
 "Show welcome again" = "Показать приветствие снова";
 "ShortcutCycle is running in your menu bar" = "ShortcutCycle работает в строке меню";
 "Look for the ShortcutCycle icon in the top-right menu bar whenever you want to open the app again." = "Ищите значок ShortcutCycle в строке меню в правом верхнем углу, когда захотите снова открыть приложение.";
-"Create your first group here, or turn on Open at Login so ShortcutCycle is always ready." = "Создайте здесь первую группу или включите «Открывать при входе», чтобы ShortcutCycle всегда был готов к работе.";
