+++ ShortcutCycle/ShortcutCycle/Resources/tr.lproj/Localizable.strings
@@ -146,4 +146,3 @@
 "Show welcome again" = "Karşılama mesajını tekrar göster";
 "ShortcutCycle is running in your menu bar" = "ShortcutCycle menü çubuğunda çalışıyor";
 "Look for the ShortcutCycle icon in the top-right menu bar whenever you want to open the app again." = "Uygulamayı tekrar açmak istediğinizde sağ üstteki menü çubuğunda ShortcutCycle simgesini arayın.";
-"Create your first group here, or turn on Open at Login so ShortcutCycle is always ready." = "Buradan ilk grubunuzu oluşturun veya ShortcutCycle'ın her zaman hazır olması için 'Girişte Aç' seçeneğini etkinleştirin.";
