+++ ShortcutCycle/ShortcutCycle/Resources/zh-Hans.lproj/Localizable.strings
@@ -147,4 +147,3 @@
 "Show welcome again" = "再次显示欢迎信息";
 "ShortcutCycle is running in your menu bar" = "ShortcutCycle 正在菜单栏中运行";
 "Look for the ShortcutCycle icon in the top-right menu bar whenever you want to open the app again." = "每当您想再次打开应用时，请在右上角菜单栏中找到 ShortcutCycle 图标。";
-"Create your first group here, or turn on Open at Login so ShortcutCycle is always ready." = "在此处创建您的第一个分组，或开启\u300c登录时打开\u300d让 ShortcutCycle 始终就绪。";
