+++ ShortcutCycle/ShortcutCycle/Resources/zh-Hant.lproj/Localizable.strings
@@ -148,4 +148,3 @@
 "Show welcome again" = "再次顯示歡迎訊息";
 "ShortcutCycle is running in your menu bar" = "ShortcutCycle 正在選單列中執行";
 "Look for the ShortcutCycle icon in the top-right menu bar whenever you want to open the app again." = "每當您想再次開啟應用程式時，請在右上角的選單列中找到 ShortcutCycle 圖示。";
-"Create your first group here, or turn on Open at Login so ShortcutCycle is always ready." = "在此處建立您的第一個群組，或開啟「登入時開啟」讓 ShortcutCycle 隨時就緒。";
