+++ ShortcutCycle/ShortcutCycle/Services/SettingsWindowLifecycleCoordinator.swift
